Move NamedOutput
