+++ src/lib.rs
@@ -235,17 +235,19 @@ use std::collections::HashMap;
 use std::ffi::{OsStr, OsString};
 use std::io::Write;
 use std::process::Command;
-use std::process::ExitStatus;
 use std::process::Output;
 use std::sync::LazyLock;
 
 mod cmd_error;
 mod command;
 mod exit_status;
+mod named_output;
 mod which_problem;
 
 pub use cmd_error::CmdError;
 pub use exit_status::ExitStatusFromCode;
+pub use named_output::NamedOutput;
+pub use named_output::OutputWithName;
 #[cfg(feature = "which_problem")]
 pub use which_problem::map_which_problem;
 
@@ -453,10 +455,7 @@ pub trait CommandWithName {
         self.mut_cmd()
             .output()
             .map_err(|io_error| CmdError::SystemError(name.clone(), io_error))
-            .map(|output| NamedOutput {
-                name: name.clone(),
-                output,
-            })
+            .map(|output| output.named(name.clone()))
             .and_then(NamedOutput::nonzero_captured)
     }
 
@@ -505,10 +504,7 @@ pub trait CommandWithName {
 
         output_and_write_streams(cmd, stdout_write, stderr_write)
             .map_err(|io_error| CmdError::SystemError(name.clone(), io_error))
-            .map(|output| NamedOutput {
-                name: name.clone(),
-                output,
-            })
+            .map(|output| output.named(name.clone()))
             .and_then(NamedOutput::nonzero_streamed)
     }
 }
@@ -608,160 +604,6 @@ impl CommandWithName for &mut NamedCommand<'_> {
 
     fn mut_cmd(&mut self) -> &mut Command {
         self.command
-    }
-}
-
-/// Extension trait for [`Output`] to generate [`NamedOutput`]
-///
-/// The primary use case is exercising a function that takes [`NamedOutput`] in its arguments in a test.
-///
-/// # Examples
-///
-/// ```
-/// use fun_run::OutputWithName;
-///
-/// let output = std::process::Output {
-///     status: std::process::ExitStatus::default(),
-///     stdout: Vec::new(),
-///     stderr: Vec::new()
-/// };
-///
-/// let named: fun_run::NamedOutput = output.named("exit 0");
-/// assert_eq!(String::from("exit 0"), named.name());
-/// ```
-///
-/// For generating an [`Output`] with a non-zero status on Unix you can use [`ExitStatusFromCode::from_code`],
-/// which builds the [`ExitStatus`] from a plain exit code without you having to bit-shift the raw wait status yourself:
-///
-/// ```
-/// use fun_run::{OutputWithName, ExitStatusFromCode};
-///
-/// let output = std::process::Output {
-///     status: std::process::ExitStatus::from_code(42),
-///     stdout: Vec::new(),
-///     stderr: Vec::new()
-/// };
-///
-/// assert_eq!(42, output.status.code().unwrap());
-///
-/// let named: fun_run::NamedOutput = output.named("exit 42");
-/// let result = named.nonzero_captured();
-/// assert!(result.is_err());
-/// ```
-pub trait OutputWithName {
-    #[must_use]
-    fn named(self, s: impl AsRef<str>) -> NamedOutput;
-}
-
-impl OutputWithName for Output {
-    fn named(self, s: impl AsRef<str>) -> NamedOutput {
-        NamedOutput {
-            name: s.as_ref().to_string(),
-            output: self,
-        }
-    }
-}
-
-/// Holds an [`Output`] of a command's execution along with its "name"
-///
-/// When paired with [`CmdError`] a `Result<NamedOutput, CmdError>` will retain the
-/// "name" of the command regardless of success or failure.
-#[derive(Debug, Clone, PartialEq, Eq)]
-pub struct NamedOutput {
-    name: String,
-    output: Output,
-}
-
-impl NamedOutput {
-    /// Check status and convert into an error if nonzero (include output in error)
-    ///
-    /// Because the [NamedOutput] does not contain information about whether it was originally
-    /// streamed or not, use this associated function when the output has not been made
-    /// available to the user. This has the effect of showing it in the event of [CmdError].
-    ///
-    /// If the output was streamed to the user use [NamedOutput::nonzero_streamed]
-    ///
-    /// # Errors
-    ///
-    /// Returns an error if the status is not zero
-    pub fn nonzero_captured(self) -> Result<NamedOutput, CmdError> {
-        nonzero_captured(self.name, self.output)
-    }
-
-    /// Check status and convert into an error if nonzero (hide output in error)
-    ///
-    /// Because the [NamedOutput] does not contain information about whether it was originally
-    /// streamed or not, use this associated function when the output has was streamed to the user.
-    /// This has the effect of hiding the output in the event of [CmdError] to prevent including
-    /// duplicate information twice.
-    ///
-    /// If the output was not streamed to the user use [NamedOutput::nonzero_captured]
-    ///
-    /// # Errors
-    ///
-    /// Returns an error if the status is not zero
-    pub fn nonzero_streamed(self) -> Result<NamedOutput, CmdError> {
-        nonzero_streamed(self.name, self.output)
-    }
-
-    /// Return the [`ExitStatus`] of the output
-    #[must_use]
-    pub fn status(&self) -> &ExitStatus {
-        &self.output.status
-    }
-
-    /// Return raw stdout
-    #[must_use]
-    pub fn stdout(&self) -> &Vec<u8> {
-        &self.output.stdout
-    }
-
-    /// Return raw stderr
-    #[must_use]
-    pub fn stderr(&self) -> &Vec<u8> {
-        &self.output.stderr
-    }
-
-    /// Return lossy stdout as a String
-    #[must_use]
-    pub fn stdout_lossy(&self) -> String {
-        String::from_utf8_lossy(&self.output.stdout).to_string()
-    }
-
-    /// Return lossy stderr as a String
-    #[must_use]
-    pub fn stderr_lossy(&self) -> String {
-        String::from_utf8_lossy(&self.output.stderr).to_string()
-    }
-
-    /// Return name of the command that was run
-    #[must_use]
-    pub fn name(&self) -> String {
-        self.name.clone()
-    }
-
-    /// Return reference of the original [Output]
-    #[must_use]
-    pub fn output(&self) -> &Output {
-        &self.output
-    }
-}
-
-impl AsRef<Output> for NamedOutput {
-    fn as_ref(&self) -> &Output {
-        &self.output
-    }
-}
-
-impl<'a> From<&'a NamedOutput> for &'a Output {
-    fn from(value: &'a NamedOutput) -> Self {
-        &value.output
-    }
-}
-
-impl From<NamedOutput> for Output {
-    fn from(value: NamedOutput) -> Self {
-        value.output
     }
 }
 
@@ -930,12 +772,9 @@ pub fn on_system_error(name: String, error: std::io::Error) -> CmdError {
 pub fn nonzero_streamed(name: String, output: impl Into<Output>) -> Result<NamedOutput, CmdError> {
     let output = output.into();
     if output.status.success() {
-        Ok(NamedOutput { name, output })
+        Ok(output.named(name))
     } else {
-        Err(CmdError::NonZeroExitAlreadyStreamed(NamedOutput {
-            name,
-            output,
-        }))
+        Err(CmdError::NonZeroExitAlreadyStreamed(output.named(name)))
     }
 }
 
@@ -952,18 +791,14 @@ pub fn nonzero_streamed(name: String, output: impl Into<Output>) -> Result<Named
 pub fn nonzero_captured(name: String, output: impl Into<Output>) -> Result<NamedOutput, CmdError> {
     let output = output.into();
     if output.status.success() {
-        Ok(NamedOutput { name, output })
+        Ok(output.named(name))
     } else {
-        Err(CmdError::NonZeroExitNotStreamed(NamedOutput {
-            name,
-            output,
-        }))
+        Err(CmdError::NonZeroExitNotStreamed(output.named(name)))
     }
 }
 
 #[cfg(test)]
 mod tests {
-    use super::*;
 
     // Nightly CI greps for this test name; update ci.yml if you rename it.
     #[test]
@@ -976,13 +811,5 @@ mod tests {
             resolved.get(OsStr::new("FUN_RUN_TEST_VAR")),
             Some(&OsString::from("1"))
         );
-    }
-
-    #[test]
-    fn test_status_from_code() {
-        for code in 0..=255 {
-            let status = ExitStatus::from_code(code);
-            assert_eq!(Some(code as i32), status.code());
-        }
     }
 }
